move `createStepMenu` into visual editor
